Algorithm plays based on the shortest path to solve the board
