+++ src/Components/BFSBoard/index.js
@@ -1,121 +1,48 @@
 import React from 'react'
 import Board from '../Board/index'
 import NeonText from '../NeonText/index'
+import { toggleAdjCells } from '../../utils/board'
 
-const BFS = ({ initialBoard, rows, cols, setHasBFSWon }) => {
+class BFSBoard extends React.Component {
 
-  const [board, setBoard] = React.useState(initialBoard)
+  state = {
+    board: JSON.parse(JSON.stringify(this.props.initialBoard)),
+    counter: 0
+  }
 
-  const MAX = 70000
-  let N = rows
+  myLoop = () => {
+    setTimeout(() => {
+      const newBoard = toggleAdjCells(this.state.board, this.props.rows, this.props.cols, this.props.cheatSet[this.state.counter][0] - 1, this.props.cheatSet[this.state.counter][1] - 1)
+      console.log(newBoard)
+      this.setState({ board: newBoard })
+      const hasWon = this.state.board.every(arr => arr.every(val => !val))
+      hasWon && setTimeout(() => this.props.setHasBFSWon(true), 250)
 
-  let matrix = Array(10).fill(0).map(() => Array(10).fill(0))
-  let parent = Array.from({ length: MAX }).map((_, i) => i)
-  let moves = Array(MAX).fill(-1)
-
-  let position = []
-  let solve = []
-
-  function fillMatrix() {
-    for (let x = 1; x <= N; x++) {
-      for (let y = 1; y <= N; y++) {
-        matrix[x][y] = board[x - 1][y - 1]
+      this.setState({ counter: this.state.counter + 1 })
+      // console.log(this.props.cheatSet.lentgh)
+      if (this.state.counter < this.props.cheatSet.length) {
+        this.myLoop();
       }
-    }
+    }, 1000)
   }
 
-  function matrixToHash(N) {
-    let state = 0
-    for (let x = 1; x <= N; x++) {
-      for (let y = 1; y <= N; y++) {
-        state <<= 1
-        state |= matrix[x][y]
-      }
-    }
-    return state
+  constructor(props) {
+    super(props)
+    console.log(props.initialBoard)
+
+    this.myLoop()
+
   }
 
-  function hashToMatrix(N, state) {
-    for (let x = N; x >= 1; x--) {
-      for (let y = N; y >= 1; y--) {
-        matrix[x][y] = state & 1
-        state >>= 1
-      }
-    }
+  render() {
+
+    return (
+      <div className="col-spacing">
+        <NeonText second="BFS" />
+        <Board board={this.state.board} rows={this.props.rows} cols={this.props.cols} handleCellClick={null} />
+      </div>
+    )
   }
-
-  function play(x, y) {
-    let dirs = [[0, -1], [-1, 0], [0, 0], [1, 0], [0, 1]]
-    dirs.map(pos => matrix[x + pos[0]][y + pos[1]] ^= 1)
-  }
-
-  function bfs(N, s0) {
-
-    let queue = []
-    queue.push(s0)
-    let state = -1
-
-    while (queue.length > 0) {
-
-      state = queue.shift()
-      let m = moves[state]
-
-      if (state === 0) break
-
-      hashToMatrix(N, state)
-
-      for (let x = 1; x <= N; x++) {
-        for (let y = 1; y <= N; y++) {
-          play(x, y)
-
-          let s = matrixToHash(N)
-
-          if (moves[s] === -1) {
-            moves[s] = m + 1
-            parent[s] = state
-            position[s] = [x, y]
-            queue.push(s)
-          }
-
-          play(x, y)
-        }
-      }
-    }
-    return state
-  }
-
-  function backtrack(s) {
-    if (s === parent[s]) return
-    backtrack(parent[s])
-    solve.push([position[s][0], position[s][1]])
-  }
-
-
-  const handleCellClick = () => {
-    console.log('oi')
-    // condição vitoria - > setHasBFSWon(true) depois de um timeout
-  }
-
-  fillMatrix()
-
-  const s = matrixToHash(N)
-  moves[s] = 0
-  const ans = bfs(N, s)
-
-  if (ans !== 0)
-    alert('Impossible!')
-
-  if (moves[ans])
-    backtrack(ans)
-
-  console.log(solve)
-
-  return (
-    <div className="col-spacing">
-      <NeonText second="BFS" />
-      <Board board={board} rows={rows} cols={cols} handleCellClick={handleCellClick} />
-    </div>
-  )
 }
 
-export default BFS
+export default BFSBoard
